+++ services/src/main/java/org/keycloak/protocol/saml/mappers/HardcodedAttributeBasedOnIdPMapper.java
@@ -13,7 +13,7 @@ public class HardcodedAttributeBasedOnIdPMapper extends AbstractSAMLProtocolMapp
     private static final List<ProviderConfigProperty> configProperties = new ArrayList<ProviderConfigProperty>();
 
     static {
-        ProviderConfigProperty property = new ProviderConfigProperty();
+        ProviderConfigProperty property;
 
         property = new ProviderConfigProperty();
         property.setName(ProtocolMapperUtils.IDP_ALIAS);
@@ -24,6 +24,7 @@ public class HardcodedAttributeBasedOnIdPMapper extends AbstractSAMLProtocolMapp
         property.setDefaultValue("");
         configProperties.add(property);
 
+        property = new ProviderConfigProperty();
         property.setName(ProtocolMapperUtils.ATTRIBUTE_VALUE);
         property.setLabel(ProtocolMapperUtils.ATTRIBUTE_CONDITIONAL_VALUES_LABEL);
         property.setHelpText(ProtocolMapperUtils.ATTRIBUTE_CONDITIONAL_VALUES_HELP_TEXT);